【CW】关联WMS的数据维护

+++ intramirror-root/intramirror-order-service/intramirror-order-service-api/src/main/java/com/intramirror/order/api/service/IOrderService.java
@@ -84,7 +84,7 @@ public interface IOrderService {
      * @param status
      * @return
      */
-    List<Map<String, Object>> getOrderListByStatusAndContainerId(int containerId, int status, Long vendorId);
+    List<Map<String, Object>> getOrderListByStatusAndContainerId(long containerId, int status, Long vendorId);
